How to make a reusable ActionSheet in SwiftUI | Bootcamp #33 - Updated.
1. Added the ActionSheet with multiple button-actions.
